// src/lib/enums/community.ts
export enum PermissionState {
    Allowed = "allowed",
    Denied = "denied",
    Unset = "unset",
}

export enum PermissionCategory {
    Administration = "administration",
    Channels = "channels",
    Messages = "messages",
    Files = "files",
    Media = "media",
    Members = "members",
    Forums = "forums",
    Bots = "bots",
    Announcements = "announcements",
}